+++ include/ambition/Ambition_polity.hh
@@ -1,7 +1,7 @@
 /*
  * Seven Kingdoms: Ambition
  *
- * Copyright 2025 Tim Sviridov
+ * Copyright 2025–26 Tim Sviridov
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -44,7 +44,7 @@ class UnitCaravan;
 namespace Ambition {
 
 namespace Coordinates {
-struct Point;
+struct Rectangle;
 }
 
 
@@ -84,7 +84,7 @@ public:
   );
 
   short findNearest7kaaMarket(
-    const Coordinates::Point location
+    const Coordinates::Rectangle& location
   ) const;
 
   std::vector<short> idleCaravans(