+++ src/ambition/Ambition_coordinates.cc
@@ -1,7 +1,7 @@
 /*
  * Seven Kingdoms: Ambition
  *
- * Copyright 2025 Tim Sviridov
+ * Copyright 2025–26 Tim Sviridov
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -48,39 +48,6 @@ Interval Point::asInterval(
   return { .x = x, .y = y };
 }
 
-Point Point::from7kaaCoordinates(
-  const _7kaaCoordinates& _7kaaCoordinates
-) {
-  return (
-    Interval {
-      .x = (_7kaaCoordinates.x - ORIGIN_7KAA_COORDINATES.x),
-      .y = (ORIGIN_7KAA_COORDINATES.y - _7kaaCoordinates.y),
-    } * SCALING_FACTOR + _7KAA_OFFSET
-  ).asCoordinates();
-}
-
-_7kaaCoordinates Point::to7kaaCoordinates(
-) const {
-  const auto point = (*this - _7KAA_OFFSET).asInterval() / SCALING_FACTOR;
-  const Point temp = {
-    .x = ORIGIN_7KAA_COORDINATES.x + point.x,
-    .y = ORIGIN_7KAA_COORDINATES.y - point.y,
-  };
-
-  if (temp.x > std::numeric_limits<short>::max()
-    || temp.x < std::numeric_limits<short>::min()
-    || temp.y > std::numeric_limits<short>::max()
-    || temp.y < std::numeric_limits<short>::min()
-  ) {
-    throw std::range_error("Coordinates out of range for 7kaa.");
-  }
-
-  return {
-    .x = short(temp.x),
-    .y = short(temp.y),
-  };
-}
-
 bool Point::within(
   const Rectangle& rectangle
 ) const {
@@ -98,12 +65,6 @@ bool Point::operator==(
     return
     x == rhs.x
     && y == rhs.y;
-}
-
-bool Point::operator==(
-  const _7kaaCoordinates &rhs
-) const noexcept {
-  return *this == Point::from7kaaCoordinates(rhs);
 }
 
 Interval Point::operator-(const Point& rhs) const {
@@ -192,6 +153,24 @@ Interval& Interval::operator/=(const long long int rhs) {
 }
 
 
+Rectangle Rectangle::_7kaaTile(
+  const Point point
+) {
+  static_assert(
+    SCALING_FACTOR % 2 == 0,
+    "SCALING_FACTOR must be even in order for a 7kaa tile to be derived from"
+    " a centre Point."
+  );
+
+  /* The Point passed for getting a 7kaa tile must not be on the border between
+   * 7kaa tiles. */
+  const auto corner
+    = point - Interval{ SCALING_FACTOR / 2, SCALING_FACTOR / 2 };
+  assert(corner.x % SCALING_FACTOR != 0 && corner.y % SCALING_FACTOR != 0);
+
+  return fromPoint(corner, _7KAA_COORDINATE_STEP);
+}
+
 Rectangle Rectangle::from7kaaCoordinates(
   const _7kaaCoordinates& _7kaaCoordinates
 ) {
@@ -201,41 +180,58 @@ Rectangle Rectangle::from7kaaRectangle(
   const _7kaaCoordinates& _7kaaCoordinatesStart,
   const _7kaaCoordinates& _7kaaCoordinatesEnd
 ) {
-  return fromPoint(
-    Point::from7kaaCoordinates(_7kaaCoordinatesStart) + _7KAA_OFFSET,
-    Interval{
-      .x = std::abs(_7kaaCoordinatesEnd.x - _7kaaCoordinatesStart.x) + 1,
-      .y = -(std::abs(_7kaaCoordinatesEnd.y - _7kaaCoordinatesStart.y) + 1),
-    } * SCALING_FACTOR
-  );
+  return {
+    .start = {
+      .x = (_7kaaCoordinatesStart.x - ORIGIN_7KAA_COORDINATES.x) * SCALING_FACTOR
+        - SCALING_FACTOR / 2,
+      .y = (ORIGIN_7KAA_COORDINATES.y - _7kaaCoordinatesEnd.y) * SCALING_FACTOR
+        - SCALING_FACTOR / 2,
+    },
+    .end = {
+      .x = (_7kaaCoordinatesEnd.x - ORIGIN_7KAA_COORDINATES.x) * SCALING_FACTOR
+        + SCALING_FACTOR / 2,
+      .y = (ORIGIN_7KAA_COORDINATES.y - _7kaaCoordinatesStart.y) * SCALING_FACTOR
+        + SCALING_FACTOR / 2,
+    },
+  };
 }
 
-
-Point Rectangle::point(
-  const long int index,
-  const Interval step
+_7kaaCoordinates Rectangle::to7kaaCoordinates(
 ) const {
-  const auto low = Point {
-    std::min(start.x, end.x),
-    std::min(start.y, end.y),
-  };
-  const auto high = Point {
-    std::max(start.x, end.x),
-    std::max(start.y, end.y),
-  };
+  if (width() > SCALING_FACTOR || height() > SCALING_FACTOR) {
+    throw std::invalid_argument("Rectangle is too large for 7kaa.");
+  }
 
-  const auto stepCountX = (high.x - low.x) / step.x + 1;
+  const auto rectangle = to7kaaRectangle();
+  return {
+    .x = rectangle.x1,
+    .y = rectangle.y1,
+  };
+}
+_7kaaRectangle Rectangle::to7kaaRectangle(
+) const {
+  if (width() == 0 || height() == 0) {
+    throw std::invalid_argument(
+      "Zero-size Rectangle cannot be reliably converted to 7kaa co-ordinates."
+    );
+  }
+
+  const auto corner = topLeft();
+  const auto x1 = static_cast<short>(
+    ORIGIN_7KAA_COORDINATES.x
+    + (corner.x + SCALING_FACTOR / 2) / SCALING_FACTOR
+  );
+  const auto y1 = static_cast<short>(
+    ORIGIN_7KAA_COORDINATES.y
+    - (corner.y - SCALING_FACTOR / 2) / SCALING_FACTOR
+  );
 
   return {
-    low.x + (index % stepCountX) * step.x,
-    low.y + (index / stepCountX) * step.y,
+    .x1 = x1,
+    .y1 = y1,
+    .x2 = static_cast<short>(x1 + (width() - 1) / SCALING_FACTOR),
+    .y2 = static_cast<short>(y1 + (height() - 1) / SCALING_FACTOR),
   };
-}
-long int Rectangle::pointCount(
-  const Interval step
-) const {
-  return ((std::max(start.x, end.x) - std::min(start.x, end.x)) / step.x + 1)
-    * ((std::max(start.y, end.y) - std::min(start.y, end.y)) / step.y + 1);
 }
 
 Point Rectangle::centre(
@@ -282,6 +278,20 @@ Rectangle Rectangle::intersection(
       .y = std::min(end.y, with.end.y),
     },
   };
+}
+
+bool Rectangle::within(
+  const Rectangle& rectangle
+) const {
+  return topLeft().within(rectangle) && bottomRight().within(rectangle);
+}
+
+
+Location* get7kaaLocation(
+  const Rectangle& at
+) {
+  const auto _7kaaCoordinates = at.to7kaaCoordinates();
+  return world.get_loc(_7kaaCoordinates.x, _7kaaCoordinates.y);
 }
 
 Rectangle viewport(