+++ src/ambition/version_details.cc
@@ -45,6 +45,7 @@ std::vector<VersionUpdates> getVersionDetails(
         .heading = HEADING_CHANGED,
         .items = {
           _("Select sortied units only if the Fort is still selected, instead of overriding the selection, preventing accidentally giving orders to the wrong units in multiplayer."),
+          _("Improved the performance of building placement highlighting."),
           _("Made development slightly easier."),
         },
       },